requirements.txt
flask==3.0.0
requests==2.31.0
python-dotenv==1.0.0
werkzeug==3.0.1
markupsafe==2.1.3
jinja2==3.1.2
itsdangerous==2.1.2
click==8.1.7
aws-cdk-lib>=2.0.0
constructs>=10.0.0
aws-cdk.aws-lambda-python-alpha>=2.0.0
